Adding more detail ..
